Twitter_clone/Switter-front/requirements.txt
flask
flask_smorest
flask_restful
marshmallow
pipenv
python-dotenv
requests